requirements.txt
# Основные библиотеки
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
pillow>=9.5.0
tqdm>=4.65.0
pandas>=1.5.0

# Для скачивания данных
gdown>=4.7.0

# Для Jupyter ноутбуков
jupyter
ipykernel

# Версии для проверки (можно удалить, это комментарий)
# Python 3.8-3.11
# CUDA 11.3+ (если есть GPU)